refactor: update build system class name and test suites
